python-bridge/requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
websockets>=12.0
